bot/requirements.txt
anyio==4.4.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
python-telegram-bot==21.4
sniffio==1.3.1
matplotlib==3.9.2
yoyo-migrations==9.0.0
psycopg2==2.9.9
asyncpg==0.29.0
gspread==6.1.2